users can now use decimal values to send/delegate/stake

+++ fx/src/main/java/org/unigrid/janus/controller/SendInfo.java
@@ -4,12 +4,10 @@ package org.unigrid.janus.controller;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 public class SendInfo {
     CosmosCredentials credentials;
     String toAddress;
-    BigDecimal amountInAtom;
+    long amountInAtom;
 }
